refactor: `core`モジュールへの移植
