+++ run.py
@@ -29,8 +29,8 @@ from starlette.responses import FileResponse
 
 from voicevox_engine import __version__
 from voicevox_engine.cancellable_engine import CancellableEngine
-from voicevox_engine.core_adapter import CoreAdapter
-from voicevox_engine.core_initializer import initialize_cores
+from voicevox_engine.core.core_adapter import CoreAdapter
+from voicevox_engine.core.core_initializer import initialize_cores
 from voicevox_engine.engine_manifest import EngineManifestLoader
 from voicevox_engine.engine_manifest.EngineManifest import EngineManifest
 from voicevox_engine.library_manager import LibraryManager
