+++ test/e2e/conftest.py
@@ -5,7 +5,7 @@ from fastapi import FastAPI
 from fastapi.testclient import TestClient
 from run import generate_app
 
-from voicevox_engine.core_initializer import initialize_cores
+from voicevox_engine.core.core_initializer import initialize_cores
 from voicevox_engine.preset import PresetManager
 from voicevox_engine.setting import SettingLoader
 from voicevox_engine.tts_pipeline.tts_engine import make_tts_engines_from_cores
