+++ voicevox_engine/core/core_wrapper.py
